// src/client/features/playback/track/client-log/utils/types.ts
interface HlsFragLoadErrorInfo {
  fragUrl: string;
  levelLoadTimes: number; // In hls.js, it means how many redundant resources we have tried in current level
  fragmentRetryTimes: number; // In hls.js, it means how many frag load retries has occurred in current level
  response: Partial<ErrorEventData['response']>;
}

interface ErrorClientLogBaseInfo {
  error_code: string;
  error_message: string;
  fatal: boolean;
  reason?: string;
  // We need this property for internal error on Hls.js
  sub_error?: string;
}

export interface ErrorClientLogInfo extends ErrorClientLogBaseInfo, Partial<HlsFragLoadErrorInfo> {
  stage?: VODStage;
  features?: FeatureInfo;
  batchedErrorCount?: number;
  tvt?: number;
  track_id?: string;
  event?: string;
  originalMessage?: string;
  levelUrl?: string;
  hasMediaKeys?: boolean;
  qos_error_type?: unknown;
  qos_error_code?: unknown;
}

export interface LiveErrorClientLogInfo extends ErrorClientLogInfo {
  cdnCache?: string;
  cdnProvider?: string;
  cdnServedBy?: string;
  cdnCloudFrontId?: string;
  cdnCloudFrontPop?: string;
}